remove yesnoenum and use bool instead
