+++ tests/devices/test_dae.py
@@ -1,7 +1,7 @@
 # pyright: reportMissingParameterType=false
 
 import pytest
-from ibex_bluesky_core.devices.dae import Dae
+from ibex_bluesky_core.devices.dae import Dae, RunstateEnum
 from ophyd_async.core import get_mock_put
 
 
@@ -17,6 +17,12 @@ def test_dae_naming(dae: Dae):
     assert dae.good_uah.name == "DAE-good_uah"
 
 
+def test_dae_runstate_string_repr(dae: Dae):
+    expected = "PROCESSING"
+    dae.run_state = RunstateEnum(expected)
+    assert str(dae.run_state) == expected
+
+
 def test_explicit_dae_naming():
     dae_explicitly_named = Dae("UNITTEST:MOCK:", name="my_special_dae")
     assert dae_explicitly_named.name == "my_special_dae"
@@ -27,18 +33,6 @@ def test_dae_monitors_correct_pvs(dae: Dae):
     assert dae.good_uah.source.endswith("UNITTEST:MOCK:DAE:GOODUAH")
     assert dae.controls.begin_run.source.endswith("UNITTEST:MOCK:DAE:BEGINRUN")
     assert dae.controls.end_run.source.endswith("UNITTEST:MOCK:DAE:ENDRUN")
-
-
-async def test_dae_read_contains_intensity_and_default_keys(dae: Dae):
-    reading = await dae.read()
-
-    assert "DAE-good_uah" in reading.keys()
-
-
-async def test_dae_describe_contains_intensity_and_default_keys(dae: Dae):
-    descriptor = await dae.describe()
-
-    assert "DAE-good_uah" in descriptor.keys()
 
 
 async def test_dae_descriptor_contains_same_keys_as_reading(dae: Dae):
